Revert "Add more verbosity levels"
