+++ freqtrade/tests/test_arguments.py
@@ -5,6 +5,7 @@ Unit test file for arguments.py
 """
 
 import argparse
+import logging
 
 import pytest
 
@@ -34,7 +35,7 @@ def test_parse_args_defaults() -> None:
     args = Arguments([], '').get_parsed_arg()
     assert args.config == 'config.json'
     assert args.dynamic_whitelist is None
-    assert args.loglevel == 0
+    assert args.loglevel == logging.INFO
 
 
 def test_parse_args_config() -> None:
@@ -52,10 +53,10 @@ def test_parse_args_db_url() -> None:
 
 def test_parse_args_verbose() -> None:
     args = Arguments(['-v'], '').get_parsed_arg()
-    assert args.loglevel == 1
+    assert args.loglevel == logging.DEBUG
 
     args = Arguments(['--verbose'], '').get_parsed_arg()
-    assert args.loglevel == 1
+    assert args.loglevel == logging.DEBUG
 
 
 def test_scripts_options() -> None:
@@ -152,7 +153,7 @@ def test_parse_args_backtesting_custom() -> None:
     call_args = Arguments(args, '').get_parsed_arg()
     assert call_args.config == 'test_conf.json'
     assert call_args.live is True
-    assert call_args.loglevel == 0
+    assert call_args.loglevel == logging.INFO
     assert call_args.subparser == 'backtesting'
     assert call_args.func is not None
     assert call_args.ticker_interval == '1m'
@@ -169,7 +170,7 @@ def test_parse_args_hyperopt_custom() -> None:
     call_args = Arguments(args, '').get_parsed_arg()
     assert call_args.config == 'test_conf.json'
     assert call_args.epochs == 20
-    assert call_args.loglevel == 0
+    assert call_args.loglevel == logging.INFO
     assert call_args.subparser == 'hyperopt'
     assert call_args.spaces == ['buy']
     assert call_args.func is not None
