+++ freqtrade/tests/test_configuration.py
@@ -6,7 +6,6 @@ Unit test file for configuration.py
 import json
 from argparse import Namespace
 from copy import deepcopy
-import logging
 from unittest.mock import MagicMock
 
 import pytest
@@ -14,7 +13,7 @@ from jsonschema import ValidationError
 
 from freqtrade import OperationalException
 from freqtrade.arguments import Arguments
-from freqtrade.configuration import Configuration, set_loggers
+from freqtrade.configuration import Configuration
 from freqtrade.constants import DEFAULT_DB_DRYRUN_URL, DEFAULT_DB_PROD_URL
 from freqtrade.tests.conftest import log_has
 
@@ -407,57 +406,3 @@ def test_check_exchange(default_conf) -> None:
         match=r'.*Exchange "unknown_exchange" not supported.*'
     ):
         configuration.check_exchange(conf)
-
-
-def test_cli_verbose_with_params(default_conf, mocker, caplog) -> None:
-    """
-    Test Configuration.load_config() with cli params used
-    """
-
-    mocker.patch('freqtrade.configuration.open', mocker.mock_open(
-        read_data=json.dumps(default_conf)))
-    # Prevent setting loggers
-    mocker.patch('freqtrade.configuration.set_loggers', MagicMock)
-    arglist = ['-vvv']
-    args = Arguments(arglist, '').get_parsed_arg()
-
-    configuration = Configuration(args)
-    validated_conf = configuration.load_config()
-
-    assert validated_conf.get('loglevel') == 3
-    assert log_has('Log level set to Level 3', caplog.record_tuples)
-
-
-def test_set_loggers() -> None:
-    """
-    Test set_loggers() update the logger level for third-party libraries
-    """
-    previous_value1 = logging.getLogger('requests').level
-    previous_value2 = logging.getLogger('ccxt.base.exchange').level
-    previous_value3 = logging.getLogger('telegram').level
-
-    set_loggers()
-
-    value1 = logging.getLogger('requests').level
-    assert previous_value1 is not value1
-    assert value1 is logging.INFO
-
-    value2 = logging.getLogger('ccxt.base.exchange').level
-    assert previous_value2 is not value2
-    assert value2 is logging.INFO
-
-    value3 = logging.getLogger('telegram').level
-    assert previous_value3 is not value3
-    assert value3 is logging.INFO
-
-    set_loggers(log_level=2)
-
-    assert logging.getLogger('requests').level is logging.DEBUG
-    assert logging.getLogger('ccxt.base.exchange').level is logging.INFO
-    assert logging.getLogger('telegram').level is logging.INFO
-
-    set_loggers(log_level=3)
-
-    assert logging.getLogger('requests').level is logging.DEBUG
-    assert logging.getLogger('ccxt.base.exchange').level is logging.DEBUG
-    assert logging.getLogger('telegram').level is logging.INFO
